zeugs
git-svn-id: https://politbuero.dyndns.org/svn/repos/promotion@1080 f4953187-ecf7-0310-90f3-8d0fedb0aae5

+++ common/baseconnection.h
@@ -94,7 +94,7 @@ public slots:
     void slotNewGpsStatus(
         const quint8& gnssMode,
         const quint8& integrationMode,
-        const quint8& info,
+        const quint16& info,
         const quint8& error,
         const quint8& numSatellitesTracked,
         const quint8& lastPvtAge,
